requirements.txt
pillow
pandas
imagehash
tqdm
pyyaml
numpy
pytesseract
